requirements.txt
python-binance
pandas
numpy
python-dotenv
requests